fix: id type을 int -> Long으로 변환

See merge request s09-webmobile2-sub2/S09P12A409!13

+++ backend/lighthouse/src/main/java/com/ssafy/lighthouse/domain/common/entity/Gugun.java
@@ -14,6 +14,6 @@ import javax.persistence.OneToOne;
 @Getter
 @NoArgsConstructor
 public class Gugun extends BaseEntity {
-    private int sidoId;
+    private Long sidoId;
     private String name;
 }
